requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.10.6
python-decouple==3.0
static3==0.7.0
gunicorn==19.4.1
psycopg2==2.6.1
django-test-without-migrations==0.4
pytz
django-s3direct
django-extensions
